+++ packages/sdk38/src/index.ts
@@ -30,6 +30,7 @@ export {
   AuthExtension,
   BankBalancesResponse,
   BankExtension,
+  BaseAccount,
   BlockResponse,
   BroadcastMode,
   DistributionCommunityPoolResponse,