+++ packages/sdk38/src/lcdapi/auth.ts
@@ -2,7 +2,15 @@
 import { Coin } from "../coins";
 import { LcdClient } from "./lcdclient";
 
-export interface CosmosSdkAccount {
+/**
+ * A Cosmos SDK base account.
+ *
+ * This type describes the base account representation as returned
+ * by the Cosmos SDK 0.37–0.39 LCD API.
+ *
+ * @see https://docs.cosmos.network/master/modules/auth/02_state.html#base-account
+ */
+export interface BaseAccount {
   /** Bech32 account address */
   readonly address: string;
   readonly coins: readonly Coin[];
@@ -16,7 +24,7 @@ export interface AuthAccountsResponse {
   readonly height: string;
   readonly result: {
     readonly type: "cosmos-sdk/Account";
-    readonly value: CosmosSdkAccount;
+    readonly value: BaseAccount;
   };
 }