+++ packages/sdk38/types/index.d.ts
@@ -28,6 +28,7 @@ export {
   AuthExtension,
   BankBalancesResponse,
   BankExtension,
+  BaseAccount,
   BlockResponse,
   BroadcastMode,
   DistributionCommunityPoolResponse,