+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - @cosmjs/sdk38: Remove `Pen` type in favour of `OfflineSigner` and remove
   `Secp256k1Pen` class in favour of `Secp256k1Wallet` which takes an
   `OfflineSigner` instead of a `SigningCallback`.
+- @cosmjs/sdk38: Rename `CosmosSdkAccount` to `BaseAccount` and export the type.
 - @cosmjs/math: Add missing integer check to `Uint64.fromNumber`. Before
   `Uint64.fromNumber(1.1)` produced some result.
 - @cosmjs/sdk38: Add `SigningCosmosClient.signAndPost` as a mid-level